feat(deals): derive deal value from line items and migrate money to BigDecimal

+++ backend/src/main/java/ooo/klae/connex/backend/dto/DealRiskCurrencySummaryDto.java
@@ -1,11 +1,12 @@
 package ooo.klae.connex.backend.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /** Bounded deal-risk totals for one currency. */
 public record DealRiskCurrencySummaryDto(
     String currency,
-    double value,
+    BigDecimal value,
     long count,
     long high,
     long medium,
